requirements.txt
python-telegram-bot==21.7
websockets==13.1
pandas==2.2.3
numpy==2.2.0
scipy==1.14.1
pyyaml==6.0.2
aiohttp==3.11.7
python-dotenv==1.0.1
requests==2.32.3
